% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_candygrams.R
\name{add_commentary}
\alias{add_commentary}
\title{Tacks commentary on to candygram announcement}
\usage{
add_commentary(person, number)
}
\arguments{
\item{person}{The candygram recipient}

\item{number}{How many grams they got}
}
\value{
A string (possibly blank)
}
\description{
Tacks commentary on to candygram announcement
}
