% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insults.R
\name{fetch}
\alias{fetch}
\title{Tells someone their new slang word isn't good.}
\usage{
fetch(name, slang_word)
}
\arguments{
\item{name}{The person's name}

\item{slang_word}{The proposed slang word.}
}
\value{
An insult string
}
\description{
Tells someone their new slang word isn't good.
}
