% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give-many-candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{Announces the number of candygrams for many people}
\usage{
give_many_candygrams(person, grams, extra_messages = NULL)
}
\arguments{
\item{grams}{How many grams each person got}

\item{extra_messages}{A vector of strings giving extra messages.}

\item{persons}{The candygram recipients}
}
\value{
A vector of candy gram announcements
}
\description{
Announces the number of candygrams for many people
}
