% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliments.R
\name{nice_style}
\alias{nice_style}
\title{Tells someone I like their clothing item.}
\usage{
nice_style(object, adjective = "vintage", follow_up = FALSE)
}
\arguments{
\item{follow_up}{Should the function follow up with questions?
Defaults to `FALSE`.}

\item{name}{The person's name}
}
\value{
A compliment string
}
\description{
Tells someone I like their clothing item.
}
