% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{Announces the number of candygrams for people.}
\usage{
give_many_candygrams(people, numbers, extra_messages = rep(NA, length(people)))
}
\arguments{
\item{people}{A vector of candygram recipients}

\item{numbers}{A vector of how many grams they got}

\item{extra_messages}{An optional vector giving extra commentary.}
}
\value{
A vector of candy gram announcements
}
\description{
Announces the number of candygrams for people.
}
