% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_many_candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{returns the number of candygrams for people.}
\usage{
give_many_candygrams(names, amount)
}
\arguments{
\item{names}{The names of the candygram recipients}

\item{amount}{The amount of the candygrams they got}
}
\value{
A string of the candy gram announcements
}
\description{
returns the number of candygrams for people.
}
