% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_many_candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{Announces the number of candygrams for a person.}
\usage{
give_many_candygrams(students, counts)
}
\arguments{
\item{counts}{How many grams they got}

\item{persons}{The candygram recipients}
}
\value{
A candy gram announcements in one string
}
\description{
Announces the number of candygrams for a person.
}
