% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{Announces the number of candygrams per person.}
\usage{
give_many_candygrams(students, counts)
}
\arguments{
\item{students}{The candygram recipient}

\item{counts}{How many grams they got}
}
\value{
A candy gram announcement
}
\description{
Announces the number of candygrams per person.
}
