% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_many_candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{Announces the number of candygrams for multiple people}
\usage{
give_many_candygrams(person, number, extra_message = NULL)
}
\arguments{
\item{persons}{The candygram recipients in a vector}

\item{numbers}{How many grams they got in a vector}

\item{extra_messages}{A string giving extra commentary in a vector, no message must be entered as "NULL"}
}
\value{
Candy gram announcements
}
\description{
Announces the number of candygrams for multiple people
}
