% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{Announces the number of candygrams for a person.}
\usage{
give_many_candygrams(person, number, extra_message = NULL)
}
\arguments{
\item{person}{The candygram recipient}

\item{number}{How many grams they got}

\item{extra_message}{A string giving extra commentary.}
}
\value{
A candy gram announcement
}
\description{
Announces the number of candygrams for a person.
}
