% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_many_candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{Announces the number of candygrams for a list of people}
\usage{
give_many_candygrams(students, counts, extra_message = NULL)
}
\arguments{
\item{counts}{a list of how many grams they each got}

\item{extra_message}{A string giving extra commentary.}

\item{studets}{a list of candygram recipient}
}
\value{
A list of candy gram announcement
}
\description{
Announces the number of candygrams for a list of people
}
