% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_many_candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{Gives Multiple Candygrams}
\usage{
give_many_candygrams(students, counts, extra_messages = NULL)
}
\arguments{
\item{students}{The a vector of student names}

\item{extra_messages}{Extra commentary for each candygram}

\item{number}{The number of candygrams each student received}
}
\value{
A vector of candygram announcements
}
\description{
Gives Multiple Candygrams
}
