% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give-many-candygrams.R
\name{give_many_candygrams}
\alias{give_many_candygrams}
\title{Announces the number of candygrams for a person.}
\usage{
give_many_candygrams(person, number, extra_message)
}
\arguments{
\item{person}{A vector of the candygram recipients}

\item{number}{A vector of how many grams they got}

\item{extra_message}{A vector of string giving extra commentary.}
}
\value{
Multiple candy gram announcements
}
\description{
Announces the number of candygrams for a person.
}
