% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliments.R
\name{really_pretty}
\alias{really_pretty}
\title{Tells someone they are pretty.}
\usage{
really_pretty(name, follow_up = TRUE)
}
\arguments{
\item{name}{The person's name}

\item{follow_up}{Should the function follow up with questions?
Defaults to `TRUE`.}
}
\value{
A compliment string
}
\description{
Tells someone they are pretty.
}
